Story #14130 : cleanup after rebase
