+++ ui/ui-frontend/projects/collect/src/assets/i18n/fr.json
@@ -138,39 +138,6 @@
       }
     }
   },
-  "RECLASSIFICATION": {
-    "EXECUTION_ACTION": "Reclasser des unités archivistiques",
-    "EXECUTE_RECLASSEMENT_MESSAGE": "La demande de reclassement a bien été lancée",
-    "COMMON": {
-      "TITLE": "Reclasser des unités archivistiques",
-      "ARCHIVE_WITH_NO_PARENTS": "Le dossier n'a pas de parents"
-    },
-    "FIRST_STEP": {
-      "SUBTITLE": "Action de reclassement",
-      "SELECTED_UNIT": "unité archivistique sélectionnée",
-      "SELECTED_UNIT_PLURAL": "unités archivistiques sélectionnées",
-      "CHILDS": "fils",
-      "INCLUDING_NB_FOLDERS_DOCUMENTS": "Comprenant {{ nbDocuments }} dossiers et documents",
-      "SELECT_ACTION_MESSAGE_1": "Sélectionner une action de reclassement",
-      "SELECT_ACTION_MESSAGE_2": "Action de reclassement",
-      "CURRENT_PARENT_FOLDERS": "Il y a 1 dossier parent actuel",
-      "CURRENT_PARENT_FOLDERS_PLURAL": "Il y a {{ nbParents }} dossiers parents actuels",
-      "MORE_THAN_FIVE_CURRENT_PARENTS_FOLDERS": "Il y a plus de 5 dossiers parents actuels",
-      "ALERT_REMOVED_CURRENT_PARENTS_FOLDERS": "Attention : les dossiers parents actuels seront retirés",
-      "ALERT_NO_ATTACHMENT_CURRENT_PARENTS_FOLDERS": "Attention : les unités archivistiques sélectionnées seront sans rattachement"
-    },
-    "REPLACE_STEP": {
-      "TITLE": "Déplacer vers un autre dossier",
-      "FROM_PARENT_SOURCE_MESSAGE": "Depuis le ou les dossier(s)"
-    },
-    "DELETE_STEP": {
-      "TITLE": "Retirer du dossier parent"
-    },
-    "ADD_STEP": {
-      "TITLE": "Ajouter à un dossier",
-      "ADD_TO_FOLDER": "Ajouter au dossier"
-    }
-  },
   "ACQUISITION_INFORMATION": {
     "PAYMENT": "Versement",
     "PROTOCOL": "Protocole",
