--- ui/ui-frontend/projects/vitamui-library/src/app/modules/reclassification-dialog/reclassification-dialog.component.spec.ts
@@ -1,212 +1,0 @@
-/*
- * Copyright French Prime minister Office/SGMAP/DINSIC/Vitam Program (2019-2022)
- * and the signatories of the "VITAM - Accord du Contributeur" agreement.
- *
- * contact@programmevitam.fr
- *
- * This software is a computer program whose purpose is to implement
- * implement a digital archiving front-office system for the secure and
- * efficient high volumetry VITAM solution.
- *
- * This software is governed by the CeCILL-C license under French law and
- * abiding by the rules of distribution of free software.  You can  use,
- * modify and/ or redistribute the software under the terms of the CeCILL-C
- * license as circulated by CEA, CNRS and INRIA at the following URL
- * "http://www.cecill.info".
- *
- * As a counterpart to the access to the source code and  rights to copy,
- * modify and redistribute granted by the license, users are provided only
- * with a limited warranty  and the software's author,  the holder of the
- * economic rights,  and the successive licensors  have only  limited
- * liability.
- *
- * In this respect, the user's attention is drawn to the risks associated
- * with loading,  using,  modifying and/or developing or reproducing the
- * software by the user in light of its specific status of free software,
- * that may mean  that it is complicated to manipulate,  and  that  also
- * therefore means  that it is reserved for developers  and  experienced
- * professionals having in-depth computer knowledge. Users are therefore
- * encouraged to load and test the software's suitability as regards their
- * requirements in conditions enabling the security of their systems and/or
- * data to be ensured and,  more generally, to use and operate it in the
- * same conditions as regards security.
- *
- * The fact that you are presently reading this means that you have had
- * knowledge of the CeCILL-C license and that you accept its terms.
- */
-
-import { ReclassificationDialogComponent } from './reclassification-dialog.component';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TranslateModule } from '@ngx-translate/core';
-import {
-  MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA,
-  MatLegacyDialog as MatDialog,
-  MatLegacyDialogRef as MatDialogRef,
-} from '@angular/material/legacy-dialog';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { BASE_URL, WINDOW_LOCATION } from '../injection-tokens';
-import { MatLegacySnackBarModule as MatSnackBarModule } from '@angular/material/legacy-snack-bar';
-import { LoggerModule } from '../logger';
-import { CriteriaDataType, CriteriaOperator, SearchCriteriaDto, SearchCriteriaTypeEnum } from '../models';
-import { of } from 'rxjs';
-import { ConfirmDialogService } from '../components/confirm-dialog';
-import { ReclassificationService } from '../services/reclassification.service';
-
-const matDialogRefSpy = jasmine.createSpyObj('MatDialogRef', ['close']);
-const matDialogSpy = jasmine.createSpyObj('MatDialog', ['open']);
-
-const confirmDialogServiceMock = {
-  confirm: () => of(true),
-  listenToEscapeKeyPress: () => of({}),
-  confirmBeforeClosing: () => of(),
-};
-
-const reclassificationServiceMock = {
-  reclassification: () => of({}),
-  searchArchiveUnitsByCriteria: () => of({}),
-  openSnackBarForWorkflow: () => of({}),
-  getTotalTrackHitsByCriteria: () => of({}),
-};
-
-describe('ReclassificationDialogComponent', () => {
-  let component: ReclassificationDialogComponent;
-  let fixture: ComponentFixture<ReclassificationDialogComponent>;
-
-  const searchCriteriaDto: SearchCriteriaDto = {
-    criteriaList: [
-      {
-        criteria: 'GUID',
-        values: [
-          {
-            value: 'aeaqaaaaaeh54ftgaamraamatl3yixiaaaaq',
-            id: 'aeaqaaaaaeh54ftgaamraamatl3yixiaaaaq',
-          },
-          {
-            value: 'aeaqaaaaaeh54ftgaay7aamac2xhibyaaaba',
-            id: 'aeaqaaaaaeh54ftgaay7aamac2xhibyaaaba',
-          },
-          {
-            value: 'aeaqaaaaaehmay6yaaqhual6ysiaariaaaba',
-            id: 'aeaqaaaaaehmay6yaaqhual6ysiaariaaaba',
-          },
-          {
-            value: 'aeaqaaaaaeh54ftgaay7aamac2xhibyaaaca',
-            id: 'aeaqaaaaaeh54ftgaay7aamac2xhibyaaaca',
-          },
-          {
-            value: 'aeaqaaaaaeh54ftgaay7aamac2xzgcyaaaba',
-            id: 'aeaqaaaaaeh54ftgaay7aamac2xzgcyaaaba',
-          },
-        ],
-        operator: CriteriaOperator.EQ,
-        category: SearchCriteriaTypeEnum.FIELDS,
-        dataType: CriteriaDataType.STRING,
-      },
-    ],
-    pageNumber: 2,
-    size: 20,
-    language: 'fr',
-    trackTotalHits: true,
-  };
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ReclassificationDialogComponent],
-      imports: [TranslateModule.forRoot(), HttpClientTestingModule, MatSnackBarModule, LoggerModule.forRoot()],
-      providers: [
-        { provide: MatDialog, useValue: matDialogSpy },
-        { provide: MatDialogRef, useValue: matDialogRefSpy },
-        { provide: BASE_URL, useValue: '/fake-api' },
-        { provide: WINDOW_LOCATION, useValue: window.location },
-        {
-          provide: MAT_DIALOG_DATA,
-          useValue: {
-            appName: 'COLLECT',
-            itemSelected: 25,
-            reclassificationCriteria: searchCriteriaDto,
-            accessContract: 'ContratTNR',
-            tenantIdentifier: 2,
-            transactionId: '1234567890',
-            selectedItemCountKnown: true,
-            archiveUnitGuidSelected: 'erer545ddfd87f5dfdf1d2fes1df2sdfs5er4e5r',
-            archiveUnitAllunitup: [],
-          },
-        },
-        { provide: ConfirmDialogService, useValue: confirmDialogServiceMock },
-        { provide: ReclassificationService, useValue: reclassificationServiceMock },
-      ],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(ReclassificationDialogComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-
-  it('should call reclassification of reclassificationService', () => {
-    // Given
-    spyOn(reclassificationServiceMock, 'reclassification').and.callThrough();
-
-    // When
-    component.onSubmit();
-
-    // Then
-    expect(reclassificationServiceMock.reclassification).toHaveBeenCalled();
-  });
-
-  it('items Selected should be grather than 0 ', () => {
-    expect(component.itemSelected).toBeGreaterThan(0);
-    expect(component.itemSelected).toEqual(25);
-  });
-
-  it('Should have an app Name ', () => {
-    expect(component.data.appName).toBeDefined();
-    expect(component.data.appName).not.toBeNull();
-    expect(component.data.appName).toEqual('COLLECT');
-  });
-
-  it('Should have a tenant identifier ', () => {
-    expect(component.data.tenantIdentifier).toBeDefined();
-    expect(component.data.tenantIdentifier).not.toBeNull();
-    expect(component.data.tenantIdentifier).toEqual(2);
-  });
-
-  it('Should have a transactionId ', () => {
-    expect(component.data.transactionId).toBeDefined();
-    expect(component.data.transactionId).not.toBeNull();
-    expect(component.data.transactionId).toEqual('1234567890');
-  });
-
-  it('should call searchArchiveUnitsByCriteria of archiveService', () => {
-    // Given
-    spyOn(reclassificationServiceMock, 'searchArchiveUnitsByCriteria').and.callThrough();
-
-    // When
-    component.calculateChilds();
-
-    // Then
-    expect(component.pendingGetChilds).toBeFalsy();
-    expect(reclassificationServiceMock.searchArchiveUnitsByCriteria).toHaveBeenCalled();
-  });
-
-  it('should call getTotalTrackHitsByCriteria of archiveService', () => {
-    // Given
-    spyOn(reclassificationServiceMock, 'getTotalTrackHitsByCriteria').and.callThrough();
-
-    // When
-    component.loadExactCount();
-
-    // Then
-    expect(reclassificationServiceMock.getTotalTrackHitsByCriteria).toHaveBeenCalled();
-  });
-
-  describe('DOM', () => {
-    it('should have 2 cdk steps', () => {
-      const elementCdkStep = fixture.nativeElement.querySelectorAll('cdk-step');
-      expect(elementCdkStep.length).toBe(2);
-    });
-  });
-});
