--- ui/ui-frontend/projects/vitamui-library/src/app/modules/reclassification-dialog/shared/vitamui-snack-bar/vitamui-snack-bar.component.scss
@@ -1,16 +1,0 @@
-:host {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: 100%;
-  width: 100%;
-}
-
-.btn-link {
-  color: white;
-  font-weight: bold;
-  margin-left: 10px;
-}
-.space-style {
-  margin-right: 30px;
-}
