+++ ui/ui-frontend/projects/vitamui-library/src/assets/shared-i18n/fr.json
@@ -5801,5 +5801,55 @@
       "MANAGEMENT": "Métadonnées de gestion",
       "OTHER": "Métadonnées système"
     }
+  },
+  "RECLASSIFICATION": {
+    "EXECUTION_ACTION": "Reclasser des unités archivistiques",
+    "ALERTE_MESSAGES": {
+      "RECLASSE_TITLE": "Reclasser des unités archivistiques",
+      "BACK_TO_SELECTION": "Revenir à la sélection",
+      "ACTION_ALERTE_TITLE": "L'action demandée n'est pas possible pour cette sélection",
+      "ACTION_ALERTE_ONE_UNIT_1": "La sélection dépasse le seuil des 10 000 unités archivistiques",
+      "ACTION_ALERTE_ONE_UNIT_2": "Veuillez revoir votre sélection"
+    },
+    "EXECUTE_RECLASSEMENT_MESSAGE": "La demande de reclassement a bien été lancée",
+    "COMMON": {
+      "TITLE": "Reclasser des unités archivistiques",
+      "ATTACHMENT_GUID_LABEL": "Identifiant technique de la position de rattachement",
+      "TARGET_GUID_TITLE_LABEL": "Intitulé de la position cible",
+      "SELECT_TARGET_GUID": "Sélectionner la position cible",
+      "SELECT_TARGET_PARENT": "Sélectionner la ou les position(s) source(s) à retirer",
+      "SELECT_TARGET_PARENT_REPLACE": "Sélectionner la ou les position(s) source(s) à conserver / à retirer",
+      "FROM_PARENT_SOURCE_MESSAGE": "Dossier(s) parent(s) actuel(s)",
+      "ARCHIVE_WITH_NO_PARENTS": "Le dossier n'a pas de parents",
+      "GUID_ALREADY_EXISTS": "L'identifiant technique n'existe pas",
+      "ATTACHMENT_ALREADY_EXISTS": "Le rattachement existe déjà",
+      "RECLASSIFICATION_TOGGLE_TREE_PLAN": "Arbres & Plans",
+      "RECLASSIFICATION_TOGGLE_UA_ID": "Identifiant d'unité archivistique"
+    },
+    "FIRST_STEP": {
+      "SUBTITLE": "Action de reclassement",
+      "SELECTED_UNIT": "unité archivistique sélectionnée",
+      "SELECTED_UNIT_PLURAL": "unités archivistiques sélectionnées",
+      "CHILDS": "fils",
+      "INCLUDING_NB_FOLDERS_DOCUMENTS": "Comprenant {{ nbDocuments }} dossiers et documents",
+      "SELECT_ACTION_MESSAGE_1": "Sélectionner une action de reclassement",
+      "SELECT_ACTION_MESSAGE_2": "Action de reclassement",
+      "CURRENT_PARENT_FOLDERS": "Il y a 1 dossier parent actuel",
+      "CURRENT_PARENT_FOLDERS_PLURAL": "Il y a {{ nbParents }} dossiers parents actuels",
+      "MORE_THAN_FIVE_CURRENT_PARENTS_FOLDERS": "Il y a plus de 5 dossiers parents actuels",
+      "ALERT_REMOVED_CURRENT_PARENTS_FOLDERS": "Attention : les dossiers parents actuels seront retirés",
+      "ALERT_NO_ATTACHMENT_CURRENT_PARENTS_FOLDERS": "Attention : les unités archivistiques sélectionnées seront sans rattachement"
+    },
+    "REPLACE_STEP": {
+      "TITLE": "Déplacer vers un autre dossier",
+      "FROM_PARENT_SOURCE_MESSAGE": "Depuis le ou les dossier(s)"
+    },
+    "DELETE_STEP": {
+      "TITLE": "Retirer du dossier parent"
+    },
+    "ADD_STEP": {
+      "TITLE": "Ajouter à un dossier",
+      "ADD_TO_FOLDER": "Ajouter au dossier"
+    }
   }
 }
